refactor(widgets): rename internal/ → _internal/, parts/ → _parts/
Prefix with _ to signal these are private implementation details.
Update all 22 import references across primitives, composites, and mixins.

Also fix bug in entry_mixin.py: TextEntryPart was incorrectly imported
from numberentry_part instead of textentry_part.
