+++ src/bootstack/widgets/_parts/__init__.py
@@ -1,0 +1,11 @@
+"""Widget parts module for composite widgets."""
+
+from bootstack.widgets._parts.textentry_part import TextEntryPart
+from bootstack.widgets._parts.numberentry_part import NumberEntryPart
+from bootstack.widgets._parts.spinnerentry_part import SpinnerEntryPart
+
+__all__ = [
+    'TextEntryPart',
+    'NumberEntryPart',
+    'SpinnerEntryPart',
+]
