+++ src/bootstack/widgets/_parts/spinnerentry_part.py
