--- src/bootstack/widgets/parts/__init__.py
@@ -1,11 +1,0 @@
-"""Widget parts module for composite widgets."""
-
-from bootstack.widgets.parts.textentry_part import TextEntryPart
-from bootstack.widgets.parts.numberentry_part import NumberEntryPart
-from bootstack.widgets.parts.spinnerentry_part import SpinnerEntryPart
-
-__all__ = [
-    'TextEntryPart',
-    'NumberEntryPart',
-    'SpinnerEntryPart',
-]
